- SQL 파일 수정
-Seller 회원가입 기능 추가

+++ src/main/java/com/project/www/domain/SellerVO.java
@@ -9,7 +9,7 @@ import lombok.*;
 @NoArgsConstructor
 public class SellerVO {
 
-    private int sellerCode;
+    private long sellerCode;
     private String sellerId;
     private String sellerPw;
     private String sellerName;
